api/requirements.txt
flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
Pillow==10.4.0
tensorflow==2.17.0